Reduce the amount of parallel state stored in `TodoListForm.jsx`, make it only send `dispatch` requests to `TodoLists.jsx`
There were quite a few subtle bugs being caused by `TodoListForm` having its own parallel copy of the todos that it wanted to update. With this change, `TodoListForm` effectively only has read access to the todo lists, and instead sends an update request to `TodoLists` whenever the user interacts with a Todo item.